+++ frontend/public/index.html
@@ -9,7 +9,7 @@
       content="web application system"
     />
 
-    <title>cdf System</title>
+    <title>upl system</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
