+++ frontend/src/components/AddProduct.js
@@ -1,4 +1,4 @@
-import React, {useState, useReducer} from 'react'
+import React, {useState} from 'react'
 import { Button, Form } from 'react-bootstrap'
 import axios from 'axios'
 import {toast} from 'react-toastify'
@@ -6,64 +6,33 @@ import {getError} from '../utils/getError'
 
 
 
-const reducer = (state, action) => {
-    switch (action.type) {
-      case 'UPDATE_REQUEST':
-        return { ...state, loadingUpdate: true };
-      case 'UPDATE_SUCCESS':
-        return { ...state, loadingUpdate: false };
-      case 'UPDATE_FAIL':
-        return { ...state, loadingUpdate: false };
-      case 'P_CREATE_REQUEST':
-          return { ...state, loadingCreate: true};
-      case 'P_CREATE_SUCCESS':
-          return { ...state, loadingCreate: false };
-      case 'P_CREATE_FAIL':
-          return { ...state, loadingCreate: false };
 
-      default:
-        return state;
-    }
-  };
 
 export default function AddProduct() {
-
-    const [{loading}, dispatch] =
-    useReducer(reducer, {
-      loading: true,
-      loadingCreate: true,
-      error: '',
-    });
-
   
     const [name, setName]= useState('');
     const [code, setProductCode] = useState('')
     const [price, setPrice] = useState('')
     const [inStock, setInStock] = useState('')
 
-
-
     const createHandler = async()=> {
         try{
-            dispatch({type: 'P_CREATE_REQUEST'})
-            await axios.post('/api/product/new',{
+           const {data} = await axios.post('/api/product/new',{
                 name,
                 code,
                 price,
                 inStock
             })
             toast.success('product added successfully');
-            dispatch({type: 'P_CREATE_SUCCESS'})
-            
+            console.log(data)
         }catch(error){
             toast.error(getError(error))
-            dispatch({type: 'P_CREATE_FAIL'})
         }
     }
 
 
   return (
-        <Form className='m-auto'>
+        <Form className='m-auto w-50'>
             <Form.Text>
                 <h1>Add New Product</h1>
             </Form.Text>
@@ -104,7 +73,7 @@ export default function AddProduct() {
                 />
             </Form.Group> 
 
-            <Button onClick={createHandler} variant='success' type='submit' className='my-4 w-100'>
+            <Button variant='success' onClick={createHandler}className='my-4 w-100'>
                 Done
             </Button>
         </Form>
