+++ frontend/src/components/Charts.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Charts() {
+  return (
+    <div>
+      Charts
+    </div>
+  )
+}
