+++ frontend/src/components/CustomerData.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function CustomerData() {
+  return (
+    <div>
+      CustomerData
+    </div>
+  )
+}
