+++ frontend/src/components/Damages.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Damages() {
+  return (
+    <div>
+      Damages
+    </div>
+  )
+}
