+++ frontend/src/components/Graphs.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Graphs() {
+  return (
+    <div>
+      Graphs
+    </div>
+  )
+}
