+++ frontend/src/components/ProductsData.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function ProductsData() {
+  return (
+    <div>
+      ProductsData
+    </div>
+  )
+}
