+++ frontend/src/components/SalesData.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function SalesData() {
+  return (
+    <div>
+      SalesData
+    </div>
+  )
+}
