+++ frontend/src/index.css
@@ -1,10 +1,6 @@
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+  height: 100%
 }
 
 code {
@@ -19,4 +15,20 @@ input{
   border: none;
   padding: 4px;
   border-radius: 5px
+}
+.details-image{
+  width: 12rem;
+  height: 300px
+}
+.dashboard-nav{
+  background-color: #10143E;
+  height: 100vh
+}
+.dashboard-nav .nav-link{
+  text-decoration: none;
+  color: white;
+  height: 40px;
+  font-weight: 500;
+  padding: 12px
+
 }
